feat: minimum user dashboard markup
Here, we have an element (<input type="checkbox">) you'd typically see in a form, but we have no form, or any way for the user to persist their changes.

So when the user sees the checkbox and interacts with it, she will expect some way of persisting those changes. But what if we implemented a proper toggle button that conveys better the state of the user profile?